Added Inventory class
