+++ src/app/app.module.ts
@@ -2,7 +2,8 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 // Classes
-import { Consonant, Gloss, Parser, Permuter, Sound, Vowel } from './class/index';
+import { Consonant, Gloss, Inventory, Parser, Permuter,
+Sound, Vowel } from './class/index';
 
 // Components
 import { AppComponent } from './app.component';
@@ -12,7 +13,7 @@ import { InputComponent, IpaComponent, OutputComponent } from './component/index
 import { ConsonantFilterPipe, VowelFilterPipe } from './pipe/index';
 
 // Services
-import { IpaService, InventoryService, TranscriptionService } from './service/index';
+import { IpaService, TranscriptionService } from './service/index';
 
 @NgModule({
   declarations: [
@@ -28,7 +29,7 @@ import { IpaService, InventoryService, TranscriptionService } from './service/in
   ],
   providers: [
     IpaService,
-    InventoryService,
+    Inventory,
     Parser,
     Permuter,
     TranscriptionService
