+++ src/app/class/index.ts
@@ -1,5 +1,6 @@
 export { Consonant } from './consonant';
 export { Gloss } from './gloss';
+export { Inventory } from './inventory';
 export { Parser } from './parser';
 export { Permuter } from './permuter';
 export { Sound } from './sound';
