+++ src/app/service/index.ts
@@ -1,3 +1,2 @@
 export { IpaService } from './ipa.service';
-export { InventoryService } from './inventory.service';
 export { TranscriptionService } from './transcription.service';
